dependencies(sdt)!: move to an extension model for multivariate stats
