+++ src/SpeciesDistributionToolkit.jl
@@ -12,9 +12,6 @@ using MakieCore
 import StatsBase
 import OffsetArrays
 
-import MultivariateStats
-import StatsAPI
-
 # We make ample use of re-export
 using Reexport
 
@@ -36,9 +33,6 @@ include("integrations/gbif_phylopic.jl")
 
 # Plotting
 include("external/Makie.jl")
-
-# Plotting
-include("external/MultivariateStats.jl")
 
 # Functions for IO
 include("io/geotiff.jl")
